feat(pixel-gtm): update tag to properly send events to Apiary
This PR makes changes to the base GTM pixel tag to ensure it functions similarly to the manual pixel.

Since `sendPixel` is a `GET` and our backend requires a `POST` request, this version injects a `pixel-support.js` script to handle the POST request.

In addition, a new `_bhp` profile cookie is automatically created as a `UUID`.

And finally, after the event is sent to Apiary, it also sends the event to the GTM tag queue. All events are prefixed with `bpx:` like `bpx:pageview`.